requirements.txt
anthropic==0.42.0
streamlit==1.41.1
pandas==2.2.3
python-dotenv==1.0.1
Pillow==11.1.0
pdf2image
python-docx
python-quickbooks
intuit-oauth
SQLAlchemy>=2.0.36
watchdog>=4.0.0
